src/requirements.txt
functions-framework==3.4.0
Flask~=2.3.2
python-telegram-bot==13.7

github==1.2.7
